已成功解决 Clang 构建的报错和警告问题：
构建配置修复
libc++/libstdc++ 冲突：移除了 CMakePresets.json 中 Clang 预设的 -stdlib=libc++ 标志，改用与 Conan profile 一致的 libstdc++
Conan Profile 更新：添加 tools.build:compiler_executables 配置确保使用正确的 Clang 编译器
代码修复
std::format 兼容性：将 error.cpp 中的 std::format 替换为 fmt::format，解决 Clang + libstdc++ 下缺少 <format> 的问题
整数精度警告：修复 quality_compressor.cpp 中的 -Wimplicit-int-conversion
未使用变量/捕获：修复 block_compressor.cpp、verify_command.cpp、compress_command.cpp、decompress_command.cpp 中的警告
符号转换警告：修复 reorder_map.cpp 中 zigzag 解码的符号转换

+++ src/algo/quality_compressor.cpp
@@ -217,7 +217,7 @@ private:
     }
 
     void writeBit(int bit) {
-        bitBuffer_ = (bitBuffer_ << 1) | bit;
+        bitBuffer_ = static_cast<std::uint8_t>((bitBuffer_ << 1) | bit);
         bitCount_++;
 
         if (bitCount_ == 8) {
